Fixing circular dependency between audioHeader.ts <-> xingHeader.ts
