+++ src/mpeg/audioHeader.ts
@@ -1,44 +1,11 @@
-import {IAudioCodec, MediaTypes} from "../iCodec";
-import {Guards} from "../utils";
-import {File} from "../file";
-import {ByteVector} from "../byteVector";
-import {CorruptFileError} from "../errors";
 import XingHeader from "./xingHeader";
 import VbriHeader from "./vbriHeader";
-
-/**
- * Indicates the MPEG version of a file or stream.
- */
-export enum MpegVersion {
-    /** Unknown version */
-    Unknown = -1,
-
-    /** MPEG-1 */
-    Version1 = 0,
-
-    /** MPEG-2 */
-    Version2 = 1,
-
-    /** MPEG-2.5 */
-    Version25 = 2
-}
-
-/**
- * Indicates the MPEG audio channel mode of a file or stream.
- */
-export enum ChannelMode {
-    /** Stereo */
-    Stereo = 0,
-
-    /** Joint Stereo */
-    JointStereo = 1,
-
-    /** Dual Channel Mono */
-    DualChannel = 2,
-
-    /** Single Channel Mono */
-    SingleChannel= 3
-}
+import {ByteVector} from "../byteVector";
+import {CorruptFileError} from "../errors";
+import {File} from "../file";
+import {IAudioCodec, MediaTypes} from "../iCodec";
+import {ChannelMode, MpegVersion} from "./mpegEnums";
+import {Guards} from "../utils";
 
 /**
  * Provides information about an MPEG audio stream.
